refactor: restructure FastAPI backend
