+++ backend/app/services/connection.py
@@ -1,0 +1,27 @@
+from typing import Dict, List
+from fastapi import WebSocket
+import json
+
+class ConnectionManager:
+    def __init__(self):
+        self.active_connections: Dict[str, List[WebSocket]] = {}
+
+    async def connect(self, websocket: WebSocket, meeting_id: str):
+        await websocket.accept()
+        if meeting_id not in self.active_connections:
+            self.active_connections[meeting_id] = []
+        self.active_connections[meeting_id].append(websocket)
+
+    def disconnect(self, websocket: WebSocket, meeting_id: str):
+        if meeting_id in self.active_connections:
+            self.active_connections[meeting_id].remove(websocket)
+
+    async def send_to_meeting(self, message: dict, meeting_id: str):
+        if meeting_id in self.active_connections:
+            for connection in self.active_connections[meeting_id]:
+                try:
+                    await connection.send_text(json.dumps(message))
+                except:
+                    pass
+
+manager = ConnectionManager()
