+++ backend/app/services/locks.py
@@ -1,0 +1,4 @@
+import asyncio
+
+vote_locks = {}
+lock_manager = asyncio.Lock()
